Revert "Feature/#178 userservice outof controller"
